Changed field type for salary and odoe.
Originally, these fields where integers on the project detail form, but
were pre-populated with 0.00 by default.  This would result in an error
when the form was saved if the values were not changed to an integer
value.  The fields where changed to decimal numbers with a maximum of
two digits.

+++ pjtk2/forms.py
@@ -594,9 +594,8 @@ class ProjectForm(forms.ModelForm):
         widget=Select(choices=Project.FUNDING_CHOICES),
         required=False)
 
-    salary = forms.IntegerField(required=False)
-
-    odoe = forms.IntegerField(required=False)
+    salary = forms.DecimalField(required=False, decimal_places=2)
+    odoe = forms.DecimalField(required=False, decimal_places=2)
 
     master_database = forms.ModelChoiceField(
         label="Master Database:",
